+++ libs/zzz/assets-data/src/AssetsData_gen.json
@@ -299,7 +299,7 @@
       "select": "IconRoleSelect42.png",
       "interknot": "IconInterKnotRole0042.png"
     },
-    "Astra": {
+    "AstraYao": {
       "full": "IconRole36.png",
       "circle": "IconRoleCircle36.png",
       "trap": "IconRoleGeneral36.png",
