+++ libs/zzz/assets/src/gen/chars/AstraYao/circle.png
