+++ libs/zzz/assets/src/gen/chars/AstraYao/full.png
