+++ libs/zzz/assets/src/gen/chars/AstraYao/index.ts
