+++ libs/zzz/assets/src/gen/chars/AstraYao/interknot.png
