+++ libs/zzz/assets/src/gen/chars/AstraYao/select.png
