+++ libs/zzz/assets/src/gen/chars/AstraYao/trap.png
