+++ libs/zzz/dm-localization/assets/locales/en/charNames_gen.json
@@ -26,7 +26,7 @@
   "Seth": "Seth",
   "Piper": "Piper",
   "Hugo": "Hugo",
-  "Astra": "Astra Yao",
+  "AstraYao": "Astra Yao",
   "Evelyn": "Evelyn",
   "Vivian": "Vivian",
   "Pulchra": "Pulchra",
