+++ libs/zzz/dm-localization/assets/locales/en/char_AstraYao_gen.json
