+++ libs/zzz/dm/src/dm/character/characterIdMap.json
@@ -21,12 +21,12 @@
   "1211": "Rina",
   "1221": "Yanagi",
   "1241": "ZhuYuan",
-  "1251": "QingYi",
+  "1251": "Qingyi",
   "1261": "Jane",
   "1271": "Seth",
   "1281": "Piper",
   "1291": "Hugo",
-  "1311": "Astra",
+  "1311": "AstraYao",
   "1321": "Evelyn",
   "1331": "Vivian",
   "1351": "Pulchra",
