+++ libs/zzz/dm/src/dm/character/characterKeys.json
@@ -1,7 +1,7 @@
 [
   "Anby",
   "Anton",
-  "Astra",
+  "AstraYao",
   "Ben",
   "Billy",
   "Burnice",
@@ -22,7 +22,7 @@
   "Nicole",
   "Piper",
   "Pulchra",
-  "QingYi",
+  "Qingyi",
   "Rina",
   "Seth",
   "Soldier0Anby",
