+++ libs/zzz/formula-ui/src/char/sheets/AstraYao.tsx
@@ -3,7 +3,7 @@ import { buffs, conditionals } from '@genshin-optimizer/zzz/formula'
 import { trans } from '../../util'
 import { createBaseSheet } from '../sheetUtil'
 
-const key: CharacterKey = 'Astra'
+const key: CharacterKey = 'AstraYao'
 // TODO: Cleanup
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 //@ts-ignore
