+++ libs/zzz/formula/src/meta.ts
@@ -30,16 +30,20 @@ export const conditionals = {
       max: 2,
     },
   },
-  Astra: {
-    boolConditional: { sheet: 'Astra', name: 'boolConditional', type: 'bool' },
+  AstraYao: {
+    boolConditional: {
+      sheet: 'AstraYao',
+      name: 'boolConditional',
+      type: 'bool',
+    },
     listConditional: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'listConditional',
       type: 'list',
       list: ['val1', 'val2'],
     },
     numConditional: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'numConditional',
       type: 'num',
       int_only: true,
@@ -2711,675 +2715,675 @@ export const formulas = {
       },
     },
   },
-  Astra: {
+  AstraYao: {
     standardDmgInst: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'standardDmgInst',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         name: 'standardDmgInst',
       },
     },
     anomalyDmgInst: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'anomalyDmgInst',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomalyDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'anomaly',
         name: 'anomalyDmgInst',
       },
     },
     anomalyBuildupInst: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'anomalyBuildupInst',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         name: 'anomalyBuildupInst',
       },
     },
     dazeInst: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'dazeInst',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         name: 'dazeInst',
       },
     },
     BasicAttackCapriccio_0_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackCapriccio_0_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackCapriccio_0_dmg',
       },
     },
     BasicAttackCapriccio_0_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackCapriccio_0_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackCapriccio_0_daze',
       },
     },
     BasicAttackCapriccio_0_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackCapriccio_0_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackCapriccio_0_anomBuildup',
       },
     },
     BasicAttackCapriccio_1_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackCapriccio_1_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackCapriccio_1_dmg',
       },
     },
     BasicAttackCapriccio_1_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackCapriccio_1_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackCapriccio_1_daze',
       },
     },
     BasicAttackCapriccio_1_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackCapriccio_1_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackCapriccio_1_anomBuildup',
       },
     },
     BasicAttackCapriccio_2_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackCapriccio_2_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackCapriccio_2_dmg',
       },
     },
     BasicAttackCapriccio_2_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackCapriccio_2_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackCapriccio_2_daze',
       },
     },
     BasicAttackCapriccio_2_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackCapriccio_2_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackCapriccio_2_anomBuildup',
       },
     },
     BasicAttackCapriccio_3_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackCapriccio_3_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackCapriccio_3_dmg',
       },
     },
     BasicAttackCapriccio_3_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackCapriccio_3_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackCapriccio_3_daze',
       },
     },
     BasicAttackCapriccio_3_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackCapriccio_3_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackCapriccio_3_anomBuildup',
       },
     },
     BasicAttackInterlude_0_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackInterlude_0_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackInterlude_0_dmg',
       },
     },
     BasicAttackInterlude_0_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackInterlude_0_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackInterlude_0_daze',
       },
     },
     BasicAttackInterlude_0_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackInterlude_0_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackInterlude_0_anomBuildup',
       },
     },
     BasicAttackChorus_0_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackChorus_0_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackChorus_0_dmg',
       },
     },
     BasicAttackChorus_0_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackChorus_0_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackChorus_0_daze',
       },
     },
     BasicAttackChorus_0_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackChorus_0_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackChorus_0_anomBuildup',
       },
     },
     BasicAttackFinale_0_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackFinale_0_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackFinale_0_dmg',
       },
     },
     BasicAttackFinale_0_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackFinale_0_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackFinale_0_daze',
       },
     },
     BasicAttackFinale_0_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'BasicAttackFinale_0_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'basic',
         name: 'BasicAttackFinale_0_anomBuildup',
       },
     },
     DashAttackLunarEclipseMelody_0_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'DashAttackLunarEclipseMelody_0_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'dash',
         name: 'DashAttackLunarEclipseMelody_0_dmg',
       },
     },
     DashAttackLunarEclipseMelody_0_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'DashAttackLunarEclipseMelody_0_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'dash',
         name: 'DashAttackLunarEclipseMelody_0_daze',
       },
     },
     DashAttackLunarEclipseMelody_0_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'DashAttackLunarEclipseMelody_0_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'dash',
         name: 'DashAttackLunarEclipseMelody_0_anomBuildup',
       },
     },
     DodgeCounterUmbrellaWaltz_0_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'DodgeCounterUmbrellaWaltz_0_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'dodgeCounter',
         name: 'DodgeCounterUmbrellaWaltz_0_dmg',
       },
     },
     DodgeCounterUmbrellaWaltz_0_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'DodgeCounterUmbrellaWaltz_0_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'dodgeCounter',
         name: 'DodgeCounterUmbrellaWaltz_0_daze',
       },
     },
     DodgeCounterUmbrellaWaltz_0_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'DodgeCounterUmbrellaWaltz_0_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'dodgeCounter',
         name: 'DodgeCounterUmbrellaWaltz_0_anomBuildup',
       },
     },
     SpecialAttackWindchimesOaths_0_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'SpecialAttackWindchimesOaths_0_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'special',
         name: 'SpecialAttackWindchimesOaths_0_dmg',
       },
     },
     SpecialAttackWindchimesOaths_0_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'SpecialAttackWindchimesOaths_0_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'special',
         name: 'SpecialAttackWindchimesOaths_0_daze',
       },
     },
     SpecialAttackWindchimesOaths_0_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'SpecialAttackWindchimesOaths_0_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'special',
         name: 'SpecialAttackWindchimesOaths_0_anomBuildup',
       },
     },
     Chord_0_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'Chord_0_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'exSpecial',
         name: 'Chord_0_dmg',
       },
     },
     Chord_0_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'Chord_0_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'exSpecial',
         name: 'Chord_0_daze',
       },
     },
     Chord_0_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'Chord_0_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'exSpecial',
         name: 'Chord_0_anomBuildup',
       },
     },
     Chord_1_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'Chord_1_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'exSpecial',
         name: 'Chord_1_dmg',
       },
     },
     Chord_1_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'Chord_1_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'exSpecial',
         name: 'Chord_1_daze',
       },
     },
     Chord_1_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'Chord_1_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'exSpecial',
         name: 'Chord_1_anomBuildup',
       },
     },
     ChainAttackTipsyConcerto_0_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'ChainAttackTipsyConcerto_0_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'chain',
         name: 'ChainAttackTipsyConcerto_0_dmg',
       },
     },
     ChainAttackTipsyConcerto_0_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'ChainAttackTipsyConcerto_0_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'chain',
         name: 'ChainAttackTipsyConcerto_0_daze',
       },
     },
     ChainAttackTipsyConcerto_0_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'ChainAttackTipsyConcerto_0_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'chain',
         name: 'ChainAttackTipsyConcerto_0_anomBuildup',
       },
     },
     UltimateFantasianSonata_0_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'UltimateFantasianSonata_0_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'ult',
         name: 'UltimateFantasianSonata_0_dmg',
       },
     },
     UltimateFantasianSonata_0_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'UltimateFantasianSonata_0_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'ult',
         name: 'UltimateFantasianSonata_0_daze',
       },
     },
     UltimateFantasianSonata_0_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'UltimateFantasianSonata_0_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'ult',
         name: 'UltimateFantasianSonata_0_anomBuildup',
       },
     },
     QuickAssistOneLuminousSky_0_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'QuickAssistOneLuminousSky_0_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'quickAssist',
         name: 'QuickAssistOneLuminousSky_0_dmg',
       },
     },
     QuickAssistOneLuminousSky_0_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'QuickAssistOneLuminousSky_0_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'quickAssist',
         name: 'QuickAssistOneLuminousSky_0_daze',
       },
     },
     QuickAssistOneLuminousSky_0_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'QuickAssistOneLuminousSky_0_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'quickAssist',
         name: 'QuickAssistOneLuminousSky_0_anomBuildup',
       },
     },
     AssistFollowUpThreeLifetimesOfFate_0_dmg: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'AssistFollowUpThreeLifetimesOfFate_0_dmg',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'standardDmg',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'assistFollowUp',
         name: 'AssistFollowUpThreeLifetimesOfFate_0_dmg',
       },
     },
     AssistFollowUpThreeLifetimesOfFate_0_daze: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'AssistFollowUpThreeLifetimesOfFate_0_daze',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'dazeBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'assistFollowUp',
         name: 'AssistFollowUpThreeLifetimesOfFate_0_daze',
       },
     },
     AssistFollowUpThreeLifetimesOfFate_0_anomBuildup: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'AssistFollowUpThreeLifetimesOfFate_0_anomBuildup',
       tag: {
         et: 'own',
         qt: 'formula',
         q: 'anomBuildup',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         attribute: 'ether',
         damageType1: 'assistFollowUp',
         name: 'AssistFollowUpThreeLifetimesOfFate_0_anomBuildup',
@@ -29974,37 +29978,37 @@ export const buffs = {
       },
     },
   },
-  Astra: {
+  AstraYao: {
     m6_dmg_: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'm6_dmg_',
       tag: {
         et: 'display',
         qt: 'combat',
         q: 'common_dmg_',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         name: 'm6_dmg_',
       },
     },
     team_dmg_: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'team_dmg_',
       tag: {
         et: 'display',
         qt: 'combat',
         q: 'common_dmg_',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         name: 'team_dmg_',
       },
     },
     enemy_defRed_: {
-      sheet: 'Astra',
+      sheet: 'AstraYao',
       name: 'enemy_defRed_',
       tag: {
         et: 'display',
         qt: 'common',
         q: 'defRed_',
-        sheet: 'Astra',
+        sheet: 'AstraYao',
         name: 'enemy_defRed_',
       },
     },
