+++ libs/zzz/stats/src/allStat_gen.json
@@ -12104,7 +12104,7 @@
         ]
       ]
     },
-    "Astra": {
+    "AstraYao": {
       "id": "1311",
       "rarity": "S",
       "attribute": "ether",
